Implements linear conflict
