+++ src/game/suggest.go
@@ -2,7 +2,6 @@ package game
 
 import (
 	"errors"
-	"math"
 	"reflect"
 	"sort"
 	"time"
@@ -16,62 +15,6 @@ type Node struct {
 }
 
 const SUGGEST_DURATION time.Duration = 1
-
-func CountMisplacedTiles(grid Grid, grid2 Grid) int {
-	sum := 0
-	size := len(grid)
-	y := 0
-	for y < size {
-		x := 0
-		for x < size {
-			if grid[y][x] != grid2[y][x] {
-				sum++
-			}
-			x++
-		}
-		y++
-	}
-	return sum
-}
-
-func Taxicab(grid Grid, grid2 Grid) int {
-	sum := 0
-	size := len(grid)
-	y := 0
-
-	for y < size {
-		x := 0
-		for x < size {
-			if grid[y][x] != grid2[y][x] {
-				expectedPos, _ := FindTileByValue(grid2, grid[y][x])
-				sum += int(math.Abs(float64(y-int(expectedPos.Y))) + math.Abs(float64(x-int(expectedPos.X))))
-			}
-			x++
-		}
-		y++
-	}
-	return sum
-}
-
-func TaxicabWithValues(grid Grid, grid2 Grid) int {
-	sum := 0
-	size := len(grid)
-	y := 0
-
-	for y < size {
-		x := 0
-		for x < size {
-			if grid[y][x] != grid2[y][x] {
-				expectedPos, _ := FindTileByValue(grid2, grid[y][x])
-				sum += int(math.Abs(float64(y-int(expectedPos.Y))) + math.Abs(float64(x-int(expectedPos.X))))
-				sum += size*size - int(grid[y][x])
-			}
-			x++
-		}
-		y++
-	}
-	return sum
-}
 
 func CompareTwoNodesByCost(n1 Node, n2 Node) int {
 	if n1.Cost < n2.Cost {
@@ -110,58 +53,12 @@ func SortListWithHeuristic(list []Node) []Node {
 	return listCopied
 }
 
-func AddToPriorityList(list []Node, node Node) []Node {
-	list = append(list, node)
-	sort.Slice(list[:], func(i, j int) bool {
-		return list[i].Heuristic < list[j].Heuristic
-	})
-	return list
-}
-
-func RemoveFromPriorityList(list []Node) ([]Node, Node) {
-	node := list[0]
-	return append(list[:0], list[1:]...), node
-}
-
-func SolvePuzzle(shuffledGrid Grid, solvedGrid Grid, timeout chan bool) ([]Coords, error) {
-	var coords []Coords
-
-	node := Node{0, TaxicabWithValues(shuffledGrid, solvedGrid), shuffledGrid, coords}
-	var openList []Node
-	openList = AddToPriorityList(openList, node)
-
-	timedOut := false
-	for len(openList) > 0 && !timedOut {
-		select {
-		case timedOut = <-timeout:
-		default:
-			openList, node = RemoveFromPriorityList(openList)
-
-			if reflect.DeepEqual(node.Grid, solvedGrid) {
-				return node.Moves, nil
-			}
-
-			possibleMoves, _ := ListMovableTiles(node.Grid)
-			for _, coords := range possibleMoves {
-				neighborGrid, _ := Move(node.Grid, coords)
-				neighborNode := Node{node.Cost + 1, node.Cost + 1 + Taxicab(neighborGrid, solvedGrid), neighborGrid, append(node.Moves, coords)}
-
-				openList = AddToPriorityList(openList, neighborNode)
-			}
-		}
-	}
-	if timedOut {
-		return make([]Coords, 0), errors.New("The solver has been stopped by a timeout")
-	}
-	return make([]Coords, 0), errors.New("No solution found by the deep puzzle algorithm")
-}
-
 func BuildNeighborList(node Node, solvedGrid Grid, possibleMoves []Coords) []Node {
 	var subList []Node
 
 	for _, coords := range possibleMoves {
 		neighborGrid, _ := Move(node.Grid, coords)
-		heuristic := TaxicabWithValues(neighborGrid, solvedGrid)
+		heuristic := TaxicabWithLinearConflict(neighborGrid, solvedGrid)
 		neighborNode := Node{node.Cost + 1 + heuristic, heuristic, neighborGrid, append(node.Moves, coords)}
 		subList = append(subList, neighborNode)
 	}
@@ -198,7 +95,7 @@ func findResolvedGridInList(list []Node) (Node, error) {
 }
 
 func buildInitialList(shuffledGrid Grid, solvedGrid Grid, possibleMoves []Coords) [][]Node {
-	initialNode := Node{0, TaxicabWithValues(shuffledGrid, solvedGrid), shuffledGrid, make([]Coords, 0)}
+	initialNode := Node{0, TaxicabWithLinearConflict(shuffledGrid, solvedGrid), shuffledGrid, make([]Coords, 0)}
 	list := make([][]Node, 0)
 	sublist := BuildNeighborList(initialNode, solvedGrid, possibleMoves)
 	list = append(list, sublist)
